backend/requirements.txt
django
requests
python-dotenv
hvac
httpx
pillow
